moved native methods back to the launcher project
