+++ Source/DualityTemplates/Templates/SolutionTemplate/Source/Launchers/DualityGame/DualityGame.csproj
@@ -5,10 +5,9 @@
 	</PropertyGroup>
 
 	<ItemGroup>
-		<PackageReference Include="AdamsLair.Duality.Backend.DefaultOpenTK" Version="4.0.0-alpha3" />
-		<PackageReference Include="AdamsLair.Duality.Backend.DotNetFramework" Version="4.0.0-alpha3" />
-		<PackageReference Include="AdamsLair.Duality.Launcher" Version="4.0.0-alpha3" />
-		<PackageReference Include="AdamsLair.Duality.Plugins.Compatibility" Version="4.0.0-alpha3" />
+		<PackageReference Include="AdamsLair.Duality.Backend.DefaultOpenTK" Version="4.0.0-alpha2" />
+		<PackageReference Include="AdamsLair.Duality.Backend.DotNetFramework" Version="4.0.0-alpha2" />
+		<PackageReference Include="AdamsLair.Duality.Plugins.Compatibility" Version="4.0.0-alpha2" />
 	</ItemGroup>
 
 	<ItemGroup>
