+++ Source/DualityTemplates/Templates/SolutionTemplate/Source/Launchers/DualityGame/NativeMethods.cs
@@ -1,0 +1,11 @@
+﻿using System;
+using System.Runtime.InteropServices;
+
+namespace DualityGame
+{
+	internal static class NativeMethods
+	{
+		[DllImport("kernel32.dll")]
+		public static extern Int32 AllocConsole();
+	}
+}
