+++ Source/DualityTemplates/Templates/SolutionTemplate/Source/Launchers/DualityGame/Program.cs
@@ -1,4 +1,7 @@
 ﻿using System;
+using System.Linq;
+using Duality;
+using Duality.Launcher;
 
 namespace DualityGame
 {
@@ -7,7 +10,18 @@ namespace DualityGame
 		[STAThread]
 		public static void Main(string[] args)
 		{
+			bool isDebugging = System.Diagnostics.Debugger.IsAttached || args.Contains(DualityApp.CmdArgDebug);
+			bool isRunFromEditor = args.Contains(DualityApp.CmdArgEditor);
+			if (isDebugging || isRunFromEditor) ShowConsole();
 			DualityLauncher.Run(args);
+		}
+
+		private static bool hasConsole = false;
+		private static void ShowConsole()
+		{
+			if (hasConsole) return;
+			SafeNativeMethods.AllocConsole();
+			hasConsole = true;
 		}
 	}
 }
