+++ Source/DualityTemplates/Templates/SolutionTemplate/Source/Launchers/DualityGame/SafeNativeMethods.cs
@@ -1,11 +1,6 @@
 ﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Runtime.InteropServices;
-using System.Drawing;
 
-namespace Duality.Launcher
+namespace DualityGame
 {
 	internal static class SafeNativeMethods
 	{
