+++ Source/Launcher/NativeMethods.cs
@@ -1,9 +1,5 @@
 ﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
 using System.Runtime.InteropServices;
-using System.Drawing;
 
 namespace Duality.Launcher
 {
