+++ Source/Launcher/Program.cs
@@ -1,4 +1,5 @@
 ﻿using System;
+using System.Linq;
 
 namespace Duality.Launcher
 {
@@ -7,7 +8,18 @@ namespace Duality.Launcher
 		[STAThread]
 		public static void Main(string[] args)
 		{
+			bool isDebugging = System.Diagnostics.Debugger.IsAttached || args.Contains(DualityApp.CmdArgDebug);
+			bool isRunFromEditor = args.Contains(DualityApp.CmdArgEditor);
+			if (isDebugging || isRunFromEditor) ShowConsole();
 			DualityLauncher.Run(args);
+		}
+
+		private static bool hasConsole = false;
+		private static void ShowConsole()
+		{
+			if (hasConsole) return;
+			SafeNativeMethods.AllocConsole();
+			hasConsole = true;
 		}
 	}
 }
