+++ Source/Launcher/SafeNativeMethods.cs
@@ -1,0 +1,20 @@
+﻿using System;
+
+namespace Duality.Launcher
+{
+	internal static class SafeNativeMethods
+	{
+		public static void AllocConsole()
+		{
+			try
+			{
+				_AllocConsole();
+			}
+			catch (Exception) { }
+		}
+		private static void _AllocConsole()
+		{
+			NativeMethods.AllocConsole();
+		}
+	}
+}
